zadanie 4.2  samochody

+++ kodilla-collections/src/main/java/com/kodilla/collections/interfaces/homework/Car.java
@@ -1,7 +1,7 @@
 package com.kodilla.collections.interfaces.homework;
 
 public interface Car {
-    int getSpeed(); //pobranie aktualnej prędkości samochodu
-    int increaseSpeed(); //zwiększenie prędkości
+    double getSpeed(); //pobranie aktualnej prędkości samochodu
+    double increaseSpeed(); //zwiększenie prędkości
     int decreaseSpeed(); //zmniejszenie prędkości
 }
